[MultiverseDev] cleaning, and doc.
