+++ src/pycram/helper.py
@@ -111,3 +111,5 @@ def find_multiverse_path() -> Optional[str]:
                 return multiverse_path + 'multiverse'
 
     return None
+
+
